requirements.txt
Flask==2.3.3
tweepy==4.14.0
feedparser
python-dotenv
praw
google-api-python-client
beautifulsoup4
requests
cloudscraper
flask-cors
serpapi
google-search-results
openai